requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
Django==4.2.20
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
PyJWT==2.9.0
PyMySQL==1.1.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
